% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_odm_terms.R
\name{check_odm_cv}
\alias{check_odm_cv}
\title{Check that a term is in a ODM Controlled Vocabulary}
\usage{
check_odm_cv(term, cv)
}
\arguments{
\item{term}{(character) The user-provided term that needs to be validated}

\item{cv}{(character) The "Term" column of the corresponding CV}
}
\value{
(logical) \code{TRUE} if there is an exact match in the CV, otherwise \code{FALSE}
}
\description{
Check that a term is in a ODM Controlled Vocabulary
}
