% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_source.R
\name{create_citation}
\alias{create_citation}
\title{Create an EDI data package citation from EML hosted on the EDI Data Portal}
\usage{
create_citation(eml = eml)
}
\arguments{
\item{Citation}{(character) If anything other than NULL is passed to this function, a citation will not be generated}
}
\value{
(character) An EDI data package citation
}
\description{
Create an EDI data package citation from EML hosted on the EDI Data Portal
}
\examples{



}
