% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_values.R
\name{create_data_values}
\alias{create_data_values}
\title{Create the DataValues table}
\usage{
create_data_values(
  L0_flat = flat,
  ValueID,
  DataValue,
  ValueAccuracy = NULL,
  LocalDateTime,
  UTCOffset,
  DateTimeUTC = NULL,
  SiteCode,
  VariableCode,
  OffsetValue = NULL,
  OffsetTypeCode = NULL,
  CensorCode = "nc",
  QualifierCode = NULL,
  MethodCode,
  QualityControlLevelCode,
  NoDataValue
)
}
\arguments{
\item{L0_flat}{L0_flat (tbl_df, tbl, data.frame) The fully joined source L0 dataset, in "flat" format (see details).}

\item{ValueID}{Column in \code{L0_flat} containing the identifier assigned to each unique data value.}

\item{DataValue}{Column in \code{L0_flat} containing the numeric value of the observation.}

\item{ValueAccuracy}{Optional. Column in \code{L0_flat} containing the umeric value that describes the measurement accuracy of the data value.}

\item{LocalDateTime}{Column in \code{L0_flat} containing local date and time at which the data value was observed.}

\item{UTCOffset}{Column in \code{L0_flat} containing offset in hours from UTC time of the corresponding LocalDateTime value.}

\item{DateTimeUTC}{Column in \code{L0_flat} containing UTC date and time at which the data value was observed.}

\item{SiteCode}{Column in \code{L0_flat} containing code used by organization that collects the data to identify the site.}

\item{VariableCode}{Column in \code{L0_flat} containing code used by the organization that collects the data to identify the variable.}

\item{OffsetValue}{Optional. Column in \code{L0_flat} containing distance from a datum or control point to the point at which a data value was observed.}

\item{CensorCode}{Column in \code{L0_flat} containing text indication of whether the data value is censored. Defaults to "nc" (Not Censored).}

\item{QualifierCode}{Optional. Column in \code{L0_flat} containing a flag indicating a peculiarity with a particular data value.}

\item{MethodCode}{Column in \code{L0_flat} containing code used by the organization that collects the data to identify the Method.}

\item{QualityControlLevelCode}{Column in \code{L0_flat} containing code which identifies the level of quality control that the value has been subjected to.}

\item{NoDataValue}{Column in \code{L0_flat} containing numeric value used to encode when a data value is not available for this variable.}

\item{OffsetTypeCodeMandatory}{if OffsetValue is used. Column in \code{L0_flat} containing code used by the organization that collects the data to identify the OffsetType.}
}
\value{

}
\description{
Create the DataValues table
}
\details{
This function appends columns to the \code{L0_flat} table and returns the augmented table.

"flat" format refers to the fully joined source L0 dataset in "wide" form with the exception of the core observation variables, which are in "long" form (i.e. using the variable_name, value, unit columns of the observation table). This "flat" format is the "widest" an L1 hymetDP dataset can be consistently spread due to the frequent occurrence of L0 source datasets with > 1 core observation variable.
}
\examples{

}
