% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_qualifiers.R
\name{create_qualifiers}
\alias{create_qualifiers}
\title{Create the Qualifiers table}
\usage{
create_qualifiers(L0_flat, QualifierCode, QualifierDescription)
}
\arguments{
\item{L0_flat}{(tbl_df, tbl, data.frame) The fully joined source L0 dataset,
in "flat" format (see details).}

\item{QualifierCode}{(character) Column in \code{L0_flat} containing the code
to indicate a given data qualifier.}

\item{QualifierDescription}{(character) Column in \code{L0_flat} containing
the text of the data qualifying comment, e.g., low battery voltage on
sensor.}
}
\value{
flat <- hymet_L0_flat

Qualifiers <- hymetDP::create_qualifiers( L0_flat = flat, QualifierCode =
"QualifierCode", QualifierDescription = "QualifierDescription")

Qualifiers
}
\description{
Create the Qualifiers table
}
\details{
This function appends columns to the \code{L0_flat} table and
returns the augmented table.

"flat" format refers to the fully joined source L0 dataset in "wide" form
with the exception of the core observation variables, which are in "long"
form (i.e. using the variable_name, value, unit columns of the observation
table). This "flat" format is the "widest" an L1 hymetDP dataset can be
consistently spread due to the frequent occurrence of L0 source datasets
with > 1 core observation variable.
}
\concept{create optional tables}
