% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sites.R
\name{create_sites}
\alias{create_sites}
\title{Create the Sites table}
\usage{
create_sites(
  L0_flat = flat,
  SiteCode,
  SiteName,
  Latitude,
  Longitude,
  LatLongDatumSRSName = NULL,
  Elevation_m = NULL,
  VerticalDatum = NULL,
  LocalX = NULL,
  LocalY = NULL,
  LocalProjectionSRSName = NULL,
  PosAccuracy_m = NULL,
  State = NULL,
  County = NULL,
  Comments = NULL,
  SiteType = NULL
)
}
\arguments{
\item{L0_flat}{(tbl_df, tbl, data.frame) The fully joined source L0 dataset, in "flat" format (see details).}

\item{SiteCode}{(character) Column in \code{L0_flat} containing the user or organization-defined code that collects the data to identify the site.}

\item{SiteName}{(character) Column in \code{L0_flat} containing the full name of the sampling site.}

\item{Latitude}{(character) Column in \code{L0_flat} containing the latitude in decimal degrees.}

\item{Longitude}{(character) Column in \code{L0_flat} containing the longitude in decimal degrees.}

\item{LatLongDatumSRSName}{(character) Column in \code{L0_flat} containing the The spatial reference system of the latitude and longitude coordinates. Choose an SRSName where IsGeographic=True from \code{SpatialReferencesCV}. View possible options with \code{SpatialReferencesCV$SRSName[SpatialReferencesCV$IsGeographic]}}

\item{Elevation_m}{(character) Column in \code{L0_flat} containing the elevation of the sampling location in meters.}

\item{VerticalDatum}{(character) Column in \code{L0_flat} containing the Vertical datum of the elevation. Choose a Term from the VerticalDatum controlled vocabulary.}

\item{LocalX}{(character) Column in \code{L0_flat} containing the local projection X coordinate.}

\item{LocalY}{(character) Column in \code{L0_flat} containing the local projection Y coordinate.}

\item{LocalProjectionSRSName}{(character) Column in \code{L0_flat} containing the full text name of the spatial reference system of the local coordinates. This field is optional and is only necessary if local coordinates are given. Choose an SRSName from from \code{SpatialReferencesCV}.}

\item{PosAccuracy_m}{(character) Column in \code{L0_flat} containing the value giving the accuracy with which the positional information is specified in meters.}

\item{State}{(character) Column in \code{L0_flat} containing the name of state in which the monitoring site is located.}

\item{County}{(character) Column in \code{L0_flat} containing the name of county in which monitoring site is located.}

\item{Comments}{(character) Column in \code{L0_flat} containing the comments related to the site.}

\item{SiteType}{(character) Column in \code{L0_flat} containing the the type of site. Choose a Term from \code{SiteTypeCV}.}
}
\value{

}
\description{
Create the Sites table
}
\details{
This function appends columns to the \code{L0_flat} table and returns the augmented table.

"flat" format refers to the fully joined source L0 dataset in "wide" form with the exception of the core observation variables, which are in "long" form (i.e. using the variable_name, value, unit columns of the observation table). This "flat" format is the "widest" an L1 hymetDP dataset can be consistently spread due to the frequent occurrence of L0 source datasets with > 1 core observation variable.
}
\examples{

}
