% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sources.R
\name{create_sources}
\alias{create_sources}
\title{Create the Sources table}
\usage{
create_sources(
  L0_flat = flat,
  SourceCode,
  Organization,
  SourceDescription,
  SourceLink = NULL,
  ContactName,
  Phone,
  Email,
  Address,
  City,
  State,
  ZipCode,
  Citation
)
}
\arguments{
\item{L0_flat}{(tbl_df, tbl, data.frame) The fully joined source L0 dataset, in "flat" format (see details).}

\item{SourceCode}{(character) Column in \code{L0_flat} containing the code which identifies the organization that created the data.}

\item{Organization}{(character) Column in \code{L0_flat} containing the name of the organization that collected the data.}

\item{SourceDescription}{(character) Column in \code{L0_flat} containing the full text description of the source of the data.}

\item{SourceLink}{(character) Column in \code{L0_flat} containing the link to the original data file of the data source.}

\item{ContactName}{(character) Column in \code{L0_flat} containing the name of the contact person for the data source.}

\item{Phone}{(character) Column in \code{L0_flat} containing the phone number for the contact person.}

\item{Email}{(character) Column in \code{L0_flat} containing the email address for the contact person.}

\item{Address}{(character) Column in \code{L0_flat} containing the street address for the contact person.}

\item{City}{(character) Column in \code{L0_flat} containing the city in which the contact person is located.}

\item{State}{(character) Column in \code{L0_flat} containing the state in which the contact person is located.}

\item{ZipCode}{(character) Column in \code{L0_flat} containing the US Zip Code or country postal code.}

\item{Citation}{(character) Column in \code{L0_flat} containing the Text string that gives the citation to be used when the data from each source are referenced.}
}
\value{
(tbl_df, tbl, data.frame) The Sources table.
}
\description{
Create the Sources table
}
\details{
This function appends columns to the \code{L0_flat} table and returns the augmented table.

"flat" format refers to the fully joined source L0 dataset in "wide" form with the exception of the core observation variables, which are in "long" form (i.e. using the variable_name, value, unit columns of the observation table). This "flat" format is the "widest" an L1 hymetDP dataset can be consistently spread due to the frequent occurrence of L0 source datasets with > 1 core observation variable.#'
}
\examples{

}
