% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_USGS_hymet.R
\name{create_USGS_hymet}
\alias{create_USGS_hymet}
\title{Create a USGS hymetDP-formatted dataset}
\usage{
create_USGS_hymet(site, param, start, end)
}
\arguments{
\item{site}{(character) A USGS site code or list of site codes. Discover site codes with \code{dataRetrieval::whatNWISsites()}.}

\item{param}{(character) A USGS parameter code or list of parameter codes. Discover parameter codes with \code{dataRetrieval::whatNWISdata()}.}

\item{start}{(character) Starting datetime for data query. Format date strings as YYYY-MM-DD, datetime strings as YYYY-MM-DDThh:mm:ssZ. Format must match \code{end} parameter format.}

\item{end}{(character) Ending datetime for data query. Format date strings as YYYY-MM-DD,}
}
\value{
A list of hymetDP tables
}
\description{
Create a USGS hymetDP-formatted dataset
}
\examples{

site <- c("06879650", "50065500")
param <- c("00060")
start <- c("2020-06-01T12:30:00Z")
end <- c("2021-01-01T12:30:00Z")

usgs_hymet <- create_USGS_hymet(site, param, start, end)

}
