% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_source.R
\name{define_source}
\alias{define_source}
\title{Define a hymetDP source}
\usage{
define_source(
  L0_flat = flat,
  eml = eml,
  Organization = NULL,
  SourceDescription = NULL,
  SourceLink = NULL,
  ContactName = NULL,
  Phone = NULL,
  Email = NULL,
  Address = NULL,
  City = NULL,
  State = NULL,
  ZipCode = NULL,
  Citation = NULL
)
}
\arguments{
\item{L0_flat}{(tbl_df, tbl, data.frame) The fully joined source L0 dataset, in "flat" format (see details).}

\item{eml}{('xml_document' 'xml_node') EML metadata.}

\item{Organization}{(character) Name of the organization that collected the data.}

\item{SourceDescription}{(character) Full text description of the source of the data.}

\item{SourceLink}{(character) Optional. Full text description of the source of the data.}

\item{ContactName}{(character) Name of the contact person for the data source.}

\item{Phone}{(character) Phone number for the contact person.}

\item{Email}{(character) Email addresss for the contact person.}

\item{Address}{(character) Street address for the contact person.}

\item{City}{(character) City in which the contact person is located.}

\item{State}{(character) State in which the contact person is located. Use two letter abbreviations for US. For other countries give the full country name.}

\item{ZipCode}{(character) US Zip Code or country postal code.}

\item{Citation}{(character) Text string that gives the citation to be used when the data from each source are referenced.}
}
\value{
(tbl_df, tbl, data.frame) An augmented version of the original flat table, with all of the original columns plus additional columns for the Source information.
}
\description{
Define a hymetDP source
}
\details{
This function appends columns to the \code{L0_flat} table and returns the augmented table.

"flat" format refers to the fully joined source L0 dataset in "wide" form with the exception of the core observation variables, which are in "long" form (i.e. using the variable_name, value, unit columns of the observation table). This "flat" format is the "widest" an L1 hymetDP dataset can be consistently spread due to the frequent occurrence of L0 source datasets with > 1 core observation variable.
}
\examples{

flat <- <insert_test_flat_table_here>

flat <- define_variable(
  L0_flat = flat,
  eml = eml,
  Organization = NULL,
  SourceDescription = NULL,
  SourceLink = NULL,
  ContactName = NULL,
  Phone = NULL,
  Email = NULL,
  Address = NULL,
  City = NULL,
  State = NULL,
  ZipCode = NULL,
  Citation = NULL)

}
