% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_variable.R
\name{define_variable}
\alias{define_variable}
\title{Define a hymetDP variable}
\usage{
define_variable(
  table = "flat",
  local_variable_column = "variable_name",
  local_variable = NULL,
  variable_name = local_variable,
  variable_units = NULL,
  sample_medium = "Unknown",
  value_type = "Unknown",
  is_regular = FALSE,
  time_support = 0,
  time_units = "hour",
  data_type = "Unknown",
  general_category = "Unknown",
  no_data = -9999
)
}
\arguments{
\item{table}{(character) The fully joined source L0 dataset, in "flat" format (see details).}

\item{local_variable_column}{(character) Column in \code{table} containing the L0 variable name.}

\item{local_variable}{(character) Reference to a value in the \code{local_variable_column} from the \code{table} to which the new hymetDP variable refers.}

\item{variable_name}{(character) The CUAHSI ODM Controlled Vocabulary name for the variable that was measured, observed, modeled, etc. Defaults to the \code{local_variable} value.}

\item{variable_units}{(character) The CUAHSI ODM Controlled Vocabulary name of units of the data values associated with a variable. Defaults to a column \code{unit} from the \code{table} if left unspecified.}

\item{sample_medium}{(character) The CUAHSI ODM Controlled Vocabulary name of the medium in which the sample or observation was taken or made.}

\item{value_type}{(character) The CUAHSI ODM Controlled Vocabulary value that indicates how the data value was generated.}

\item{is_regular}{(boolean) Value indicates whether the data values are from a regularly sampled time series. Choose \code{TRUE} or \code{FALSE}.}

\item{time_support}{(numeric) Numerical value that indicates the time support (or temporal footprint) of the data values. 0 is used to indicate data values that are instantaneous. Other values indicate the time over which the data values are implicitly or explicitly averaged or aggregated. Goes along with \code{time_units} if \code{is_regular == TRUE}.}

\item{time_units}{(character) The CUAHSI ODM Controlled Vocabulary name of units of the time support. If \code{time_support == 0}, indicating an instantaneous observation, a unit needs to still be given for completeness, although it is arbitrary.}

\item{data_type}{(character) The CUAHSI ODM Controlled Vocabulary value that indicates how the value applies over a time interval.}

\item{general_category}{(character) The CUAHSI ODM Controlled Vocabulary value for general category of the data  (i.e. Hydrology).}

\item{no_data}{(numeric) Numeric value used to encode when a data value is not available for this variable. DataValues will be reformatted to match this value.}
}
\value{
(tbl_df, tbl, data.frame) An augmented version of the original flat table, with all of the original columns plus one for each of the variable values (i.e. variable_name, variable_units, etc.).
}
\description{
Define a hymetDP variable
}
\examples{

flat <- <insert_test_flat_table_here>

flat <- define_variable(
  table = "flat",
  local_variable_column = "variable_name",
  local_variable = NULL,
  variable_name = local_variable,
  variable_units = NULL,
  sample_medium = "Unknown",
  value_type = "Unknown",
  is_regular = FALSE,
  time_support = 0,
  time_units = "hour",
  data_type = "Unknown",
  general_category = "Unknown",
  no_data = -9999)

}
