% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GeneralCategoryCV}
\alias{GeneralCategoryCV}
\title{Allowed General Categories}
\format{
A data frame with 10 rows and 2 variables:
\describe{
\item{Term}{Controlled Vocabulary term}
\item{Definition}{Definition of Controlled Vocabulary term}
}
}
\source{
\url{http://his.cuahsi.org/mastercvreg/edit_cv11.aspx?tbl=GeneralCategoryCV}
}
\usage{
GeneralCategoryCV
}
\description{
A data frame representation of the CUAHSI ODM 1.1 Controlled Vocabulary for General Category
}
\keyword{datasets}
