% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_USGS_hymet.R
\name{get_usgs_sitetype}
\alias{get_usgs_sitetype}
\title{Get USGS site type}
\usage{
get_usgs_sitetype(s)
}
\arguments{
\item{s}{(character) USGS site type code \code{site_type_cd}}
}
\value{
Site type
}
\description{
Get USGS site type
}
