% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_USGS_hymet.R
\name{get_usgs_variable}
\alias{get_usgs_variable}
\title{Get USGS variable information}
\usage{
get_usgs_variable(p)
}
\arguments{
\item{p}{(character) USGS parameter code \code{param}}
}
\value{
Information associated with a USGS variable
}
\description{
Get USGS variable information
}
\keyword{internal}
