% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hymet_L0_flat}
\alias{hymet_L0_flat}
\title{Joined and flat version of EDI data package knb-lter-mcm.9003.11}
\format{
A data frame with 30,000 rows and 45 variables:
\describe{
\item{LocalDateTime}{Local datetime}
\item{UTCOffset}{Local timezone offset from UTC}
\item{DateTimeUTC}{Datetime in UTC}
\item{Variable_name}{Variable name from the L0 dataset}
\item{DataValue}{Numeric value of the observation}
\item{Qualifier}{Code of a qualifier (flag)}
\item{Unit}{Unit name from the L0 dataset}
\item{ValueID}{The data value ID}
\item{VariableCode}{The variable ID}
\item{VariableName}{The ODM CV variable name}
\item{VariableUnitsName}{The ODM CV unit name for the variable}
\item{SampleMedium}{The ODM CV sample medium name}
\item{ValueType}{The ODM CV value type}
\item{IsRegular}{Whether the values are from a regularly sampled dataset}
\item{TimeSupport}{Temporal footprint of samples}
\item{TimeUnitsName}{The ODM CV unit name for time support}
\item{DataType}{The ODM CV data type}
\item{GeneralCategory}{The ODM CV general category}
\item{NoDataValue}{The numeric value used to encode not available data}
\item{MethodCode}{The method ID}
\item{MethodDescription}{The description of the method}
\item{SiteCode}{The site ID}
\item{SiteName}{The name of the site}
\item{Latitude}{Latitude of site}
\item{Longitude}{Longitude of site}
\item{Elevation_m}{Elevation of site in meters}
\item{LatLongDatumSRSName}{Spatial reference system of latitude and longitude coordinates}
\item{SiteType}{The ODM CV site type}
\item{Organization}{Name of organization that collected the data}
\item{ContactName}{Name of contact person for the data source}
\item{Email}{Email of contact}
\item{SourceCode}{The source ID}
\item{SourceDescription}{The description of the source data}
\item{SourceLink}{URL of source data}
\item{Citation}{Citation for the source data}
\item{Phone}{Contact phone}
\item{Address}{Contact physical address}
\item{City}{Contact city}
\item{State}{Contact state if in US, otherwise full country name}
\item{ZipCode}{Contact postal code}
\item{QualityControlLevelCode}{The quality control ID}
\item{Definition}{Definition of the quality control level}
\item{Explanation}{Explanation of the quality control level}
\item{QualifierCode}{The qualifier ID}
\item{QualifierDescription}{Description of the qualifier}
}
}
\source{
\url{https://portal.edirepository.org/nis/mapbrowse?scope=knb-lter-mcm&identifier=9003&revision=11}
}
\usage{
hymet_L0_flat
}
\description{
A fully joined and flat version of EDI data package knb-lter-mcm.9003.11 (Seasonal high-frequency measurements of discharge, water temperature, and specific conductivity from Andersen Creek at H1, McMurdo Dry Valleys, Antarctica (1993-2020, ongoing)) with all relevant hymetDP L1 identifiers and content added. Use this dataset as an input to the \code{L0_flat} argument of the "create" functions.
}
\keyword{datasets}
