% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hymet_L1}
\alias{hymet_L1}
\title{The hymetDP (L1) version of EDI data package knb-lter-mcm.9003.11}
\format{
A list of:
\describe{
\item{id}{The dataset identifier}
\item{metadata}{See source url for metadata}
\item{tables}{A list of data frames, each a hymetDP table}
}
}
\source{
\code{hymet_L0_flat}
}
\usage{
hymet_L1
}
\description{
The the hymetDP (L1) formatted version of EDI data package knb-lter-mcm.9003.11 (Seasonal high-frequency measurements of discharge, water temperature, and specific conductivity from Andersen Creek at H1, McMurdo Dry Valleys, Antarctica (1993-2020, ongoing)) produced from the table \code{hymet_L0_flat}. Use this dataset as an input to data "use" functions.
}
\keyword{datasets}
