% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odm_controlled_vocabulary.R
\name{odm_controlled_vocabulary}
\alias{odm_controlled_vocabulary}
\title{View one of CUAHSI's ODM Controlled Vocabularies}
\usage{
odm_controlled_vocabulary(vocab = "Units")
}
\arguments{
\item{vocab}{(character) The controlled vocabulary to be returned.}
}
\value{
(tbl_df, tbl, data.frame) The controlled vocabulary formatted as a tibble.
}
\description{
This function returns a user-selected ODM controlled vocabulary to the console. The returned object may be useful for searching for appropriate terms.
}
\details{
Use the search field to find suitable terms for creating hymetDP tables.
}
\examples{

\dontrun{
# View the ODM VariableNameCV

View(odm_controlled_vocabulary('VariableName'))
}

}
