% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read published data}
\usage{
read_data(
  id = NULL,
  parse_datetime = TRUE,
  unique_keys = FALSE,
  from = NULL,
  format = "new"
)
}
\arguments{
\item{id}{(character) Identifier of dataset to read. Identifiers are listed in the "id" column of the \code{search_data()} output. Older versions of datasets can be read, but a warning is issued.}

\item{parse_datetime}{(logical) Parse datetime values if TRUE, otherwise return as character strings.}

\item{unique_keys}{(logical) Whether to create globally unique primary keys (and associated foreign keys). Useful in maintaining referential integrity when working with multiple datasets. If TRUE, \code{id} is appended to each table's primary key and associated foreign key. Default is FALSE.}

\item{from}{(character) Full path of file to be read (if .rds), or path to directory containing saved datasets (if .csv).}

\item{format}{(character) Format of returned object, which can be: "new" (the new implementation) or "old" (the original implementation; deprecated). In the new format, the top most level of nesting containing the "id" field has been moved to the same level as the "tables", "metadata", and "validation_issues" fields.}
}
\value{
(list) A dataset with the structure:
\itemize{
\item id - Dataset identifier
\item metadata - List of info about the dataset. NOTE: This object is underdevelopment and content may change in future releases.
\item tables - List of dataset tables as data.frames.
\item validation_issues - List of validation issues. If the dataset fails any validation checks, then descriptions of each issue are listed here.
}
}
\description{
Read published data
}
\details{
\preformatted{Validation checks are applied to each dataset ensuring it complies with the hymetDP model. A warning is issued when any validation checks fail. All datasets are returned, even if they fail validation.

Column classes are coerced to those defined in the hymetDP specification.

Validation happens each time files are read, from source APIs or local environments.
}
}
\note{
This function may not work between 01:00 - 03:00 UTC on Wednesdays due to regular maintenance of the EDI Data Repository.
}
