% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_data_package_report_hymet}
\alias{read_data_package_report_hymet}
\title{Read data package report}
\usage{
read_data_package_report_hymet(packageId, frmt = "xml", env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{frmt}{(character) Format of the returned report. Can be: "xml", "html", or "char".}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(xml_document) Data package report
}
\description{
Read data package report
}
\examples{
# Read as XML
qualityReport <- read_data_package_report_hymet("knb-lter-knz.260.4")
qualityReport

# Read as HTML
qualityReport <- read_data_package_report_hymet("knb-lter-knz.260.4", frmt = "html")
qualityReport

# Read as character
qualityReport <- read_data_package_report_hymet("knb-lter-knz.260.4", frmt = "char")
# writeLines(qualityReport, "./data/report.txt"))

}
\keyword{internal}
