% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_metadata}
\alias{read_metadata}
\title{Read metadata}
\usage{
read_metadata(packageId, env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{env}{(character) Repository environment. Can be: "production",
"staging", or "development".}
}
\value{
(xml_document) EML metadata document.

See the
\href{https://CRAN.R-project.org/package=emld}{emld} library
for more on working with EML as a list or JSON-LD. See the
\href{https://CRAN.R-project.org/package=xml2}{xml2} library
for working with EML as XML.
}
\description{
Read metadata
}
\examples{
\dontrun{

# Read metadata
eml <- read_metadata("edi.100.1")
eml
#> {xml_document}
#> <eml packageId="edi.100.1" system="https://pasta.edirepository.org"   ...
#> [1] <access authSystem="https://pasta.edirepository.org/authenticatio ...
#> [2] <dataset>\n  <alternateIdentifier system="https://doi.org">doi:10 ...
}


}
\concept{Accessing}
