% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_odm_terms.R
\name{return_close}
\alias{return_close}
\title{Return values from the CV that may be close to users' term}
\usage{
return_close(term, cv)
}
\arguments{
\item{term}{(character) The user-provided term that needs to be validated}

\item{cv}{(character) The "Term" column of the corresponding CV}
}
\value{
(list) Named list with any CV entries that contain the string
}
\description{
Return values from the CV that may be close to users' term
}
