% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_data.R
\name{search_data}
\alias{search_data}
\title{Search published data}
\usage{
search_data(
  text,
  VariableName,
  SampleMedium,
  GeneralCategory,
  SiteType,
  TimeSupport,
  starts_before,
  ends_after,
  num_years,
  area,
  boolean = "AND"
)
}
\arguments{
\item{text}{(character) Text to search for in dataset titles and abstracts. Datasets matching any exact words or phrase will be returned. Can be a regular expression as used by \code{stringr::str_detect()}. Is not case sensitive. Works with \code{boolean}.}

\item{VariableName}{(character) VariableName values to search on. VariableName values are from ODM Controlled Vocabulary.}

\item{SampleMedium}{(character) SampleMedium values to search on. SampleMedium values are from ODM Controlled Vocabulary.}

\item{GeneralCategory}{(character) GeneralCategory values to search on. GeneralCategory values are from ODM Controlled Vocabulary.}

\item{SiteType}{(character) SiteType values to search on. SiteType values are from ODM Controlled Vocabulary.}

\item{TimeSupport}{(numeric) Maximum TimeSupport value to search on. TimeSupport is analogous frequency of measurements.}

\item{starts_before}{(date) Maximum start date to filter on.}

\item{ends_after}{(date) Minimum end date to filter on.}

\item{num_years}{(numeric) Minimum and maximum number of years sampled the dataset should contain. Any datasets within this range will be returned.}

\item{area}{(numeric) Bounding coordinates within which the data should originate. Accepted values are in decimal degrees and in the order: North, East, South, West. Any datasets with overlapping areas or contained points will be returned.}

\item{boolean}{(character) Boolean operator to use when searching \code{text}, \code{VariableName}, \code{SampleMedium}, \code{GeneralCategory}, and
\code{SiteType}. Supported operators are: "AND", "OR". Default is "AND".}
}
\value{
(tbl_df, tbl, data.frame) Search results with these fields:
\itemize{
\item source - Source from which the dataset originates. Currently supported are "EDI" and "NEON".
\item id - Identifier of the dataset.
\item title - Title of the dataset.
\item abstract - Abstract of dataset.
\item years - Number of years sampled.
\item url - URL to dataset.
\item source_id - Identifier of source L0 dataset.
\item source_id_url - URL to source L0 dataset.
}
}
\description{
Search published data
}
\examples{
\dontrun{
# Empty search returns all available datasets
search_data()

# "text" searches titles, descriptions, and abstracts
search_data(text = "barometric")

# "VariableName" searches VariableName values for a match
search_data(VariableName = "Discharge")

# "SampleMedium" searches SampleMedium values for a match
search_data(SampleMedium = "Water")

# "GeneralCategory" searches GeneralCategory values for a match
search_data(GeneralCategory = "Hydrology")

# "SiteType" searches SiteType values for a match
search_data(SiteType = "Stream")

# "TimeSupport" searches TimeSupport values for a match
search_data(TimeSupport = 30)

# "starts_before" and "ends_after" can be used to filter on a time period
search_data(starts_before = '2000-01-01', ends_after = '2010-01-01')

# "num_years" searches the number of years sampled
search_data(num_years = c(10, 20))

# Use any combination of search fields to find the data you're looking for
search_data(
  text = c("stream", "river"),
  VariableName = c("Conductivity", "Discharge"),
  SampleMedium = "water",
  GeneralCategory = "hydrology"
  SiteType = "Stream"
  TimeSupport = 30,
  starts_before = "2010-01-01",
  ends_after = "2015-01-01",
  num_years = c(10, 100),
  area = c(47.1, -86.7, 42.5, -92),
  boolean = "OR")
}

}
