% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SiteTypeCV}
\alias{SiteTypeCV}
\title{Allowed Site Types}
\format{
A data frame with 62 rows and 2 variables:
\describe{
\item{Term}{Controlled Vocabulary term}
\item{Definition}{Definition of Controlled Vocabulary term}
}
}
\source{
\url{http://his.cuahsi.org/mastercvreg/edit_cv11.aspx?tbl=SiteTypeCV}
}
\usage{
SiteTypeCV
}
\description{
A data frame representation of the CUAHSI ODM 1.1 Controlled Vocabulary for Site Types
}
\keyword{datasets}
