% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SpatialReferencesCV}
\alias{SpatialReferencesCV}
\title{Allowed Spatial References}
\format{
A data frame with 343 rows and 5 variables:
\describe{
\item{SpatialReferencesID}{Integer identifier}
\item{SRSID}{Spatial Reference System identifier}
\item{SRSName}{Spatial Reference System name}
\item{IsGeographic}{Boolean}
\item{Notes}{Spatial Reference System notes}
}
}
\source{
\url{http://his.cuahsi.org/mastercvreg/edit_cv11.aspx?tbl=SpatialReferences}
}
\usage{
SpatialReferencesCV
}
\description{
A data frame representation of the CUAHSI ODM 1.1 Controlled Vocabulary for Spatial References
}
\keyword{datasets}
