% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supported_USGS_params}
\alias{supported_USGS_params}
\title{hymetDP Supported USGS parameters}
\format{
A data frame with 1 row and 11 variables:
\describe{
\item{usgs_code}{USGS Parameter Code}
\item{VariableName}{Corresponding hymetDP/ODM CV term for VariableName}
\item{VariableUnitsName}{Corresponding hymetDP/ODM CV term for VariableUnitsName}
\item{SampleMedium}{Corresponding hymetDP/ODM CV term for SampleMedium}
\item{ValueType}{Corresponding hymetDP/ODM CV term for ValueType}
\item{IsRegular}{Corresponding hymetDP/ODM CV term for IsRegular}
\item{TimeSupport}{Corresponding hymetDP/ODM CV term for TimeSupport}
\item{TimeUnitsName}{Corresponding hymetDP/ODM CV term for TimeUnitsName}
\item{DataType}{Corresponding hymetDP/ODM CV term for DataType}
\item{GeneralCategory}{Corresponding hymetDP/ODM CV term for GeneralCategory}
\item{NoDataValue}{Corresponding hymetDP/ODM CV term for NoDataValue}
}
}
\usage{
supported_USGS_params
}
\description{
The USGS NWIS parameter codes that can be reformatted as hymetDP with \code{create_USGS_hymet()}
}
\keyword{datasets}
