% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_hymet.R
\name{template_hymet}
\alias{template_hymet}
\title{Template create_hymetDP.R script}
\usage{
template_hymet(path, dir.name)
}
\arguments{
\item{path}{(character) Path to where a new directory should be created}

\item{dir.name}{(character) Name of the new directory which will contain the create_hymetDP.R script}
}
\value{
Empty create_hymetDP.R script.
}
\description{
Template create_hymetDP.R script
}
\examples{
\dontrun{
template_hymet(tempdir(), 'my_hymet')
}

}
