% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UnitsCV}
\alias{UnitsCV}
\title{Allowed Units}
\format{
A data frame with 394 rows and 4 variables:
\describe{
\item{UnitsID}{Unique unit identifier}
\item{UnitsName}{Name of the unit}
\item{UnitsType}{Type of unit}
\item{UnitsAbbreviation}{Standardized abbreviation or symbol of the unit}
}
}
\source{
\url{http://his.cuahsi.org/mastercvreg/edit_cv11.aspx?tbl=Units}
}
\usage{
UnitsCV
}
\description{
A data frame representation of the CUAHSI ODM 1.1 Controlled Vocabulary for Units
}
\keyword{datasets}
