% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_odm_terms.R
\name{validate_odm_terms}
\alias{validate_odm_terms}
\title{Validate arguments of hymetDP functions that should be using CUAHSI ODM CV terms}
\usage{
validate_odm_terms(fun.name, fun.args)
}
\arguments{
\item{fun.name}{(character) Name of function from which \code{validate_omd_terms()} is
called.}

\item{fun.args}{(named list) Arguments passed to calling function and formatted as
\code{as.list(environment())}.}
}
\description{
\preformatted{Validate input arguments to hymetDP functions that should have direct matches to terms in CUAHSI ODM CVs.
}
}
\details{
\preformatted{Validation checks are function specific.
}
}
\keyword{internal}
