% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{write_tables}
\alias{write_tables}
\title{Write tables to file}
\usage{
write_tables(
  path,
  sep = ",",
  DataValues = NULL,
  Variables = NULL,
  Methods = NULL,
  Sources = NULL,
  Sites = NULL,
  QualityControlLevels = NULL,
  SeriesCatalog = NULL,
  Qualifiers = NULL
)
}
\arguments{
\item{path}{(character) A path to the directory in which the files will be written.}

\item{sep}{(character) Field delimiter to use when writing files. Default is comma.}

\item{DataValues}{(tbl_df, tbl, data.frame) The DataValues table.}

\item{Variables}{(tbl_df, tbl, data.frame) The Variables table.}

\item{Methods}{(tbl_df, tbl, data.frame) The Methods table.}

\item{Sources}{(tbl_df, tbl, data.frame) The Sources table.}

\item{Sites}{(tbl_df, tbl, data.frame) The Sites table.}

\item{QualityControlLevels}{(tbl_df, tbl, data.frame) The QualityControlLevels table.}

\item{SeriesCatalog}{(tbl_df, tbl, data.frame) The SeriesCatalog table.}

\item{Qualifiers}{(tbl_df, tbl, data.frame) The Qualifiers table.}
}
\value{
hymetDP tables as \code{sep} delimited files
}
\description{
Write tables to file
}
