% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_source.R
\name{create_citation}
\alias{create_citation}
\title{Create an EDI data package citation from EML hosted on the EDI Data Portal}
\usage{
create_citation(eml = eml)
}
\arguments{
\item{eml}{(character) An EML document from which a valid EDI package identifier can be extracted.}
}
\value{
(character) An EDI data package citation.
}
\description{
Create an EDI data package citation from EML hosted on the EDI Data Portal
}
\examples{

}
\keyword{internal}
