% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_methods.R
\name{create_methods}
\alias{create_methods}
\title{Title}
\usage{
create_methods(
  L0_flat = flat,
  MethodCode,
  MethodDescription,
  MethodLink = NULL
)
}
\arguments{
\item{L0_flat}{(tbl_df, tbl, data.frame) The fully joined source L0 dataset, in "flat" format (see details).}

\item{MethodCode}{(character) Column in \code{L0_flat} containing the code used by the organization that collects the data to identify the Method.}

\item{MethodDescription}{(character) Column in \code{L0_flat} containing the text description of each method.}

\item{MethodLink}{(character) Optional. Column in \code{L0_flat} containing a link to additional reference material on the method. Should be a single valid URL.}
}
\value{
(tbl_df, tbl, data.frame) The Methods table.
}
\description{
Title
}
\examples{

}
