% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_series_catalog.R
\name{create_series_catalog}
\alias{create_series_catalog}
\title{Create the SeriesCatalog table}
\usage{
create_series_catalog(
  L0_flat = NULL,
  Sources = NULL,
  Methods = NULL,
  Variables = NULL,
  Sites = NULL,
  QualityControlLevels = NULL,
  DataValues = NULL
)
}
\arguments{
\item{L0_flat}{(tbl_df, tbl, data.frame) The fully joined source L0 dataset, in "flat" format (see details).}

\item{Sources}{(tbl_df, tbl, data.frame) The Sources table.}

\item{Methods}{(tbl_df, tbl, data.frame) The Methods table.}

\item{Variables}{(tbl_df, tbl, data.frame) The Variables table.}

\item{Sites}{(tbl_df, tbl, data.frame) The Sites table.}

\item{QualityControlLevels}{(tbl_df, tbl, data.frame) The QualityControlLevels table.}

\item{DataValues}{(tbl_df, tbl, data.frame) The DataValues table.}
}
\value{
(tbl_df, tbl, data.frame) The SeriesCatalog table.
}
\description{
Create the SeriesCatalog table
}
\details{
This function appends columns to the \code{L0_flat} table and returns the augmented table.

"flat" format refers to the fully joined source L0 dataset in "wide" form with the exception of the core observation variables, which are in "long" form (i.e. using the variable_name, value, unit columns of the observation table). This "flat" format is the "widest" an L1 hymetDP dataset can be consistently spread due to the frequent occurrence of L0 source datasets with > 1 core observation variable.
}
\examples{

}
