% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_method.R
\name{define_method}
\alias{define_method}
\title{Define a hymetDP method}
\usage{
define_method(
  L0_flat = flat,
  local_variable_column = "variable_name",
  local_variable = NULL,
  VariableCode = NULL,
  MethodDescription = NULL,
  MethodLink = NULL
)
}
\arguments{
\item{L0_flat}{(tbl_df, tbl, data.frame) The fully joined source L0 dataset, in "flat" format (see details).}

\item{local_variable_column}{(character) Column in \code{L0_flat} table containing the L0 variable name.}

\item{local_variable}{(character) Reference to a value (or values) in the \code{local_variable_column} from the \code{L0_flat} table to which the new hymetDP method refers.}

\item{VariableCode}{(character) The auto-generated primary key for a variable (from column \code{VariableCode}). Another way to link a method to a value (or values). Takes priority over \code{local_variable}. Provide one or multiple codes.}

\item{MethodDescription}{(character) Text description of the method.}

\item{MethodLink}{(character) Optional. Link to additional reference material on the method. Should be a single valid URL.}
}
\value{
(tbl_df, tbl, data.frame) An augmented version of the original flat table, with all of the original columns plus one additional column for the method description (or two additional columns if a method link is defined). Column name includes the auto-generated MethodCode (i.e. MethodDescription_1), which will become the the primary key in the Methods table.
}
\description{
Define a hymetDP method
}
\details{
This function appends columns to the \code{L0_flat} table and returns the augmented table.

"flat" format refers to the fully joined source L0 dataset in "wide" form with the exception of the core observation variables, which are in "long" form (i.e. using the variable_name, value, unit columns of the observation table). This "flat" format is the "widest" an L1 hymetDP dataset can be consistently spread due to the frequent occurrence of L0 source datasets with > 1 core observation variable.
}
\examples{

flat <- <insert_test_flat_table_here>

flat <- define_variable(
  L0_flat = flat,
  local_variable_column = "variable_name",
  local_variable = NULL,
  variable_code = NULL,
  method_description = NULL,
  method_link = NULL)


}
